+++ src/constants/font-weight.js
@@ -6,6 +6,10 @@ const FW_700 = css`
   font-weight: 700;
 `;
 
+const FW_600 = css`
+  font-weight: 600;
+`;
+
 const FW_400 = css`
   font-weight: 400;
 `;
@@ -16,6 +20,7 @@ const FW_100 = css`
 
 const FONT_WEIGHT = {
   FW_700,
+  FW_600,
   FW_400,
   FW_100,
 };